Extract `pycroft.lib.user_id`
